refactor: move StyledInfo to new dir
